+++ src/test/java/com/unzer/payment/business/AuthorizationTest.java
@@ -32,6 +32,7 @@ import java.time.LocalDate;
 import java.util.Currency;
 
 import static com.unzer.payment.business.BasketV1TestData.getMaxTestBasketV1;
+import static com.unzer.payment.business.Keys.*;
 import static com.unzer.payment.util.Uuid.generateUuid;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.awaitility.Awaitility.await;
@@ -71,7 +72,7 @@ public class AuthorizationTest extends AbstractPaymentTest {
 
     @Test
     public void testAuthorizeWithPaymentType() throws MalformedURLException, HttpCommunicationException {
-        Unzer unzer = getUnzer(privateKey2);LocalDate locaDateNow = LocalDate.now();
+        Unzer unzer = getUnzer(KEY_WITH_3DS);LocalDate locaDateNow = LocalDate.now();
         Card card = new Card("4444333322221111", "12/" + (locaDateNow.getYear() + 1));
         Authorization authorize = unzer.authorize(BigDecimal.ONE, Currency.getInstance("EUR"), card, new URL("https://www.unzer.com"), false);
         assertNotNull(authorize);
@@ -82,7 +83,7 @@ public class AuthorizationTest extends AbstractPaymentTest {
 
     @Test
     public void testAuthorizeReturnPaymentTypeAndCustomer() throws MalformedURLException, HttpCommunicationException {
-        Unzer unzer = getUnzer(privateKey2);LocalDate locaDateNow = LocalDate.now();
+        Unzer unzer = getUnzer(KEY_WITH_3DS);LocalDate locaDateNow = LocalDate.now();
         Card card = new Card("4444333322221111", "12/" + (locaDateNow.getYear() + 1));
         Customer customer = new Customer("Max", "Mustermann");
         Authorization authorize = unzer.authorize(BigDecimal.ONE, Currency.getInstance("EUR"), card, new URL("https://www.unzer.com"), customer, false);
@@ -115,7 +116,7 @@ public class AuthorizationTest extends AbstractPaymentTest {
 
     @Test
     public void testAuthorizeWithCustomerTypeReturnUrl() throws MalformedURLException, HttpCommunicationException {
-        Unzer unzer = getUnzer(privateKey2);LocalDate locaDateNow = LocalDate.now();
+        Unzer unzer = getUnzer(KEY_WITH_3DS);LocalDate locaDateNow = LocalDate.now();
         Card card = new Card("4444333322221111", "12/" + (locaDateNow.getYear() + 1));
         Customer customer = new Customer("Max", "Mustermann");
         Authorization authorize = unzer.authorize(BigDecimal.ONE, Currency.getInstance("EUR"), card, new URL("https://www.unzer.com"), customer, false);
@@ -234,17 +235,17 @@ public class AuthorizationTest extends AbstractPaymentTest {
             maxBasket.getBasketItems().get(i).setAmountDiscount(null);
         }
 
-        Basket basket = getUnzer(marketplacePrivatekey).createBasket(maxBasket);
+        Basket basket = getUnzer(MARKETPLACE_KEY).createBasket(maxBasket);
 
         //create card
         Card card = getPaymentTypeCard(NO_3DS_VISA_CARD_NUMBER); //do not change card number except error case
-        card = getUnzer(marketplacePrivatekey).createPaymentType(card);
+        card = getUnzer(MARKETPLACE_KEY).createPaymentType(card);
 
         //marketplace authorization
         MarketplaceAuthorization authorizeRequest = getMarketplaceAuthorization(card.getId(), null, null, null, basket.getId(), null);
         authorizeRequest.setAmount(maxBasket.getAmountTotalGross());
 
-        MarketplaceAuthorization authorize = getUnzer(marketplacePrivatekey).marketplaceAuthorize(authorizeRequest);
+        MarketplaceAuthorization authorize = getUnzer(MARKETPLACE_KEY).marketplaceAuthorize(authorizeRequest);
         assertNotNull(authorize.getId());
         assertNotNull(authorize);
         assertEquals(AbstractTransaction.Status.PENDING, authorize.getStatus());
@@ -255,7 +256,7 @@ public class AuthorizationTest extends AbstractPaymentTest {
         assertEquals(HttpStatus.SC_MOVED_TEMPORARILY, redirectStatus);
 
         //get marketplace payment
-        MarketplacePayment payment = getUnzer(marketplacePrivatekey).fetchMarketplacePayment(authorize.getPayment().getId());
+        MarketplacePayment payment = getUnzer(MARKETPLACE_KEY).fetchMarketplacePayment(authorize.getPayment().getId());
         assertNotNull(payment);
         assertNotNull(payment.getId());
         assertNotNull(payment.getAuthorizationsList());
@@ -263,7 +264,7 @@ public class AuthorizationTest extends AbstractPaymentTest {
         assertEquals(Payment.State.PENDING, payment.getPaymentState());
 
         //get marketplace authorize
-        authorize = getUnzer(marketplacePrivatekey).fetchMarketplaceAuthorization(authorize.getPayment().getId(), authorize.getId());
+        authorize = getUnzer(MARKETPLACE_KEY).fetchMarketplaceAuthorization(authorize.getPayment().getId(), authorize.getId());
         assertNotNull(authorize.getId());
         assertNotNull(authorize);
         assertEquals(AbstractTransaction.Status.SUCCESS, authorize.getStatus());