+++ src/test/java/com/unzer/payment/business/ChargeTest.java
@@ -36,6 +36,8 @@ import java.time.LocalDate;
 import java.util.Currency;
 
 import static com.unzer.payment.business.BasketV1TestData.getMaxTestBasketV1;
+import static com.unzer.payment.business.Keys.KEY_WITH_3DS;
+import static com.unzer.payment.business.Keys.MARKETPLACE_KEY;
 import static com.unzer.payment.util.Uuid.generateUuid;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -78,7 +80,7 @@ public class ChargeTest extends AbstractPaymentTest {
 
     @Test
     public void testChargeWithPaymentType() throws MalformedURLException, HttpCommunicationException {
-        Unzer unzer = getUnzer(privateKey2);LocalDate locaDateNow = LocalDate.now();
+        Unzer unzer = getUnzer(KEY_WITH_3DS);LocalDate locaDateNow = LocalDate.now();
         Card card = new Card("4444333322221111", "12/" + (locaDateNow.getYear() + 1));
         Charge charge = unzer.charge(BigDecimal.ONE, Currency.getInstance("EUR"), card, new URL("https://www.unzer.com"), false);
         assertNotNull(charge);
@@ -98,7 +100,7 @@ public class ChargeTest extends AbstractPaymentTest {
 
     @Test
     public void testChargeWithCustomerTypeReturnUrl() throws MalformedURLException, HttpCommunicationException {
-        Unzer unzer = getUnzer(privateKey2);LocalDate locaDateNow = LocalDate.now();
+        Unzer unzer = getUnzer(KEY_WITH_3DS);LocalDate locaDateNow = LocalDate.now();
         Card card = new Card("4444333322221111", "12/" + (locaDateNow.getYear() + 1));
         Customer customer = new Customer("Max", "Mustermann");
         Charge charge = unzer.charge(BigDecimal.ONE, Currency.getInstance("EUR"), card, new URL("https://www.unzer.com"), customer, false);
@@ -119,7 +121,7 @@ public class ChargeTest extends AbstractPaymentTest {
 
     @Test
     public void testChargeReturnPayment() throws MalformedURLException, HttpCommunicationException {
-        Unzer unzer = getUnzer(privateKey2);LocalDate locaDateNow = LocalDate.now();
+        Unzer unzer = getUnzer(KEY_WITH_3DS);LocalDate locaDateNow = LocalDate.now();
         Card card = new Card("4444333322221111", "12/" + (locaDateNow.getYear() + 1));
         Charge charge = unzer.charge(BigDecimal.ONE, Currency.getInstance("EUR"), card, new URL("https://www.unzer.com"), false);
         assertNotNull(charge);
@@ -233,24 +235,24 @@ public class ChargeTest extends AbstractPaymentTest {
             maxBasket.getBasketItems().get(i).setAmountDiscount(null);
         }
 
-        Basket basket = getUnzer(marketplacePrivatekey).createBasket(maxBasket);
+        Basket basket = getUnzer(MARKETPLACE_KEY).createBasket(maxBasket);
 
         //create card
         Card card = getPaymentTypeCard(NO_3DS_VISA_CARD_NUMBER); //do not change card number except error case
-        card = getUnzer(marketplacePrivatekey).createPaymentType(card);
+        card = getUnzer(MARKETPLACE_KEY).createPaymentType(card);
 
         //marketplace charge
         MarketplaceCharge chargeRequest = getMarketplaceCharge(card.getId(), null, null, null, basket.getId(), null);
         chargeRequest.setAmount(maxBasket.getAmountTotalGross());
 
-        MarketplaceCharge charge = getUnzer(marketplacePrivatekey).marketplaceCharge(chargeRequest);
+        MarketplaceCharge charge = getUnzer(MARKETPLACE_KEY).marketplaceCharge(chargeRequest);
         assertNotNull(charge.getId());
         assertNotNull(charge);
         assertEquals(AbstractTransaction.Status.PENDING, charge.getStatus());
         assertEquals(participantId_2, charge.getProcessing().getParticipantId());
 
         //get marketplace payment
-        MarketplacePayment payment = getUnzer(marketplacePrivatekey).fetchMarketplacePayment(charge.getPayment().getId());
+        MarketplacePayment payment = getUnzer(MARKETPLACE_KEY).fetchMarketplacePayment(charge.getPayment().getId());
         assertNotNull(payment);
         assertNotNull(payment.getId());
         assertNotNull(payment.getAuthorizationsList());
@@ -276,23 +278,23 @@ public class ChargeTest extends AbstractPaymentTest {
             maxBasket.getBasketItems().get(i).setAmountDiscount(null);
         }
 
-        Basket basket = getUnzer(marketplacePrivatekey).createBasket(maxBasket);
+        Basket basket = getUnzer(MARKETPLACE_KEY).createBasket(maxBasket);
 
         //create SepaDirectDebit
         SepaDirectDebit sdd = getSepaDirectDebit();
-        sdd = getUnzer(marketplacePrivatekey).createPaymentType(sdd);
+        sdd = getUnzer(MARKETPLACE_KEY).createPaymentType(sdd);
 
         //marketplace charge
         MarketplaceCharge chargeRequest = getMarketplaceCharge(sdd.getId(), null, null, null, basket.getId(), null);
         chargeRequest.setAmount(maxBasket.getAmountTotalGross());
 
-        MarketplaceCharge charge = getUnzer(marketplacePrivatekey).marketplaceCharge(chargeRequest);
+        MarketplaceCharge charge = getUnzer(MARKETPLACE_KEY).marketplaceCharge(chargeRequest);
         assertNotNull(charge.getId());
         assertNotNull(charge);
         assertEquals(AbstractTransaction.Status.SUCCESS, charge.getStatus());
 
         //get marketplace payment
-        MarketplacePayment payment = getUnzer(marketplacePrivatekey).fetchMarketplacePayment(charge.getPayment().getId());
+        MarketplacePayment payment = getUnzer(MARKETPLACE_KEY).fetchMarketplacePayment(charge.getPayment().getId());
         assertNotNull(payment);
         assertNotNull(payment.getId());
         assertNotNull(payment.getAuthorizationsList());
@@ -318,23 +320,23 @@ public class ChargeTest extends AbstractPaymentTest {
             maxBasket.getBasketItems().get(i).setAmountDiscount(null);
         }
 
-        Basket basket = getUnzer(marketplacePrivatekey).createBasket(maxBasket);
+        Basket basket = getUnzer(MARKETPLACE_KEY).createBasket(maxBasket);
 
         //create sofort
         Sofort sofort = new Sofort();
-        sofort = getUnzer(marketplacePrivatekey).createPaymentType(sofort);
+        sofort = getUnzer(MARKETPLACE_KEY).createPaymentType(sofort);
 
         //marketplace charge
         MarketplaceCharge chargeRequest = getMarketplaceCharge(sofort.getId(), null, null, null, basket.getId(), null);
         chargeRequest.setAmount(maxBasket.getAmountTotalGross());
 
-        MarketplaceCharge charge = getUnzer(marketplacePrivatekey).marketplaceCharge(chargeRequest);
+        MarketplaceCharge charge = getUnzer(MARKETPLACE_KEY).marketplaceCharge(chargeRequest);
         assertNotNull(charge.getId());
         assertNotNull(charge);
         assertEquals(AbstractTransaction.Status.PENDING, charge.getStatus());
 
         //get marketplace payment
-        MarketplacePayment payment = getUnzer(marketplacePrivatekey).fetchMarketplacePayment(charge.getPayment().getId());
+        MarketplacePayment payment = getUnzer(MARKETPLACE_KEY).fetchMarketplacePayment(charge.getPayment().getId());
         assertNotNull(payment);
         assertNotNull(payment.getId());
         assertNotNull(payment.getAuthorizationsList());
@@ -342,7 +344,7 @@ public class ChargeTest extends AbstractPaymentTest {
         assertEquals(Payment.State.PENDING, payment.getPaymentState());
 
         //get marketplace charge
-        MarketplaceCharge getCharge = getUnzer(marketplacePrivatekey).fetchMarketplaceCharge(charge.getPayment().getId(), charge.getId());
+        MarketplaceCharge getCharge = getUnzer(MARKETPLACE_KEY).fetchMarketplaceCharge(charge.getPayment().getId(), charge.getId());
         assertNotNull(getCharge);
         assertEquals(charge.getId(), getCharge.getId());
         assertEquals(AbstractTransaction.Status.PENDING, charge.getStatus());