+++ src/test/java/com/unzer/payment/business/ShipmentTest.java
@@ -17,6 +17,9 @@ package com.unzer.payment.business;
 
 
 import com.unzer.payment.PaymentException;
+import com.unzer.payment.Unzer;
+import com.unzer.payment.communication.HttpCommunicationException;
+import com.unzer.payment.paymenttypes.InvoiceSecured;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -24,13 +27,16 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 public class ShipmentTest extends AbstractPaymentTest {
     @Test
     public void testAuthorizeWithShipmentNotSameAddressWithInvoiceSecured() {
-        assertThrows(PaymentException.class, () -> getUnzer()
-                .authorize(
-                        getAuthorization(
-                                createPaymentTypeInvoiceSecured().getId(),
-                                createMaximumCustomer().getId()
-                        )
-                )
+        assertThrows(PaymentException.class, () -> {
+                    Unzer unzer = getUnzer();
+                    InvoiceSecured paymentTypeInvoiceSecured = unzer.createPaymentType(new InvoiceSecured());
+                    unzer.authorize(
+                            getAuthorization(
+                                    paymentTypeInvoiceSecured.getId(),
+                                    createMaximumCustomer().getId()
+                            )
+                    );
+                }
         );
     }