+++ src/test/java/com/unzer/payment/business/paymenttypes/CardTest.java
@@ -29,6 +29,7 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Currency;
 
+import static com.unzer.payment.business.Keys.PRIVATE_KEY_3;
 import static org.junit.jupiter.api.Assertions.*;
 
 
@@ -39,7 +40,7 @@ public class CardTest extends AbstractPaymentTest {
         assertThrows(PaymentException.class, () -> {
             Card card = new Card("4444333322221111", "03/20");
             card.setCvc("123");
-            Unzer unzer = new Unzer(new HttpClientBasedRestCommunication(), privateKey3);
+            Unzer unzer = new Unzer(new HttpClientBasedRestCommunication(), PRIVATE_KEY_3);
             unzer.createPaymentType(card);
         });
     }