+++ src/test/java/com/unzer/payment/util/ApplePayAdapterTest.java
@@ -16,38 +16,47 @@
 package com.unzer.payment.util;
 
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 
 import java.net.URISyntaxException;
+import java.util.Collections;
 
+import static com.unzer.payment.util.ApplePayAdapterUtil.doesUrlContainValidDomainName;
 import static org.junit.jupiter.api.Assertions.*;
 
-public class ApplePayAdapterTest {
 
-    @Test
-    public void ifAllParametersAreNullThrowError() {
-        assertThrows(NullPointerException.class, () -> {
-            ApplePayAdapterUtil.validateApplePayMerchant(null, null, null, null);
-        });
+@Execution(ExecutionMode.SAME_THREAD)
+public class ApplePayAdapterTest {
+    @BeforeEach
+    public void setDefaultValues() {
+        ApplePayAdapterUtil.setCustomAppleValidationUrls(ApplePayAdapterUtil.DEFAULT_VALIDATION_URLS);
     }
 
     @Test
-    public void getPlainDomainName() throws URISyntaxException {
-        String url = "https://www.unzer.com/de/";
-        String plainDomainName = ApplePayAdapterUtil.getPlainDomainName(url);
-        assertEquals("unzer.com", plainDomainName);
+    public void ifAllParametersAreNullThrowError() {
+        assertThrows(NullPointerException.class, () -> ApplePayAdapterUtil.validateApplePayMerchant(
+                null,
+                null,
+                null,
+                null));
     }
 
     @Test
     public void doesUrlsContainValidDomainName() throws URISyntaxException {
-        String validUrl1 = "https://www.apple-pay-gateway.apple.com/";
-        String validUrl2 = "https://cn-apple-pay-gateway.apple.com/";
-        String invalidUrl1 = "https://www.google.com/";
-        String invalidUrl2 = "https://www.amazon.com/";
+        assertTrue(doesUrlContainValidDomainName("https://www.apple-pay-gateway.apple.com/"));
+        assertTrue(doesUrlContainValidDomainName("https://cn-apple-pay-gateway.apple.com/"));
 
-        assertTrue(ApplePayAdapterUtil.doesUrlContainValidDomainName(validUrl1));
-        assertTrue(ApplePayAdapterUtil.doesUrlContainValidDomainName(validUrl2));
-        assertFalse(ApplePayAdapterUtil.doesUrlContainValidDomainName(invalidUrl1));
-        assertFalse(ApplePayAdapterUtil.doesUrlContainValidDomainName(invalidUrl2));
+        assertFalse(doesUrlContainValidDomainName("https://www.google.com/"));
+        assertFalse(doesUrlContainValidDomainName("https://www.amazon.com/"));
+    }
+
+    @Test
+    public void customValidationUrls() throws URISyntaxException {
+        ApplePayAdapterUtil.setCustomAppleValidationUrls(Collections.singletonList("google.com"));
+        assertTrue(doesUrlContainValidDomainName("https://www.google.com"));
+        assertFalse(doesUrlContainValidDomainName("https://www.apple-pay-gateway.apple.com/"));
     }
 }